[REXA-24] Move common code for exception in specific package
