+++ src/main/kotlin/fr/yvbro/rexa/config/RexaApiConfiguration.kt
@@ -1,6 +1,6 @@
 package fr.yvbro.rexa.config
 
-import fr.yvbro.rexa.exception.ApiErrorReporter
+import fr.yvbro.rexa.exception.common.ApiErrorReporter
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
